backend/requirements_old.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# IA (quando for implementar)
# openai==1.3.0
# google-generativeai==0.3.1
# httpx==0.25.2  # Para Groq